Add files via upload
The attributes that are in this dataset are frequency of SMS, subscription length, age group, seconds of use, status, frequency of use, and Customer Value. All of the attributes except for attribute churn is the aggregated data of the first 9 months. The objective is all about measuring why customers are leaving the company or stating whether customers will change telecommunication providers or not and to check which variables are effecting the churn value.